refactor: separate renderer settings
